add RoundDto, SeasonDto & StandingDto

+++ BackEnd-SeasonService/src/main/java/tn/iit/entity/Standing.java
@@ -4,13 +4,10 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
 
-import java.math.BigInteger;
-import java.util.List;
-
 @Data
 @Builder
 @Jacksonized
 public class Standing {
-    private BigInteger rank;
-    private BigInteger score;
+    private int rank;
+    private int score;
 }
